backend/requirements-base.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleak==2.1.1
blinker==1.9.0
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.0
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cssselect==1.4.0
csvw==3.7.0
ctranslate2==4.7.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.13
ddgs==9.11.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
duckduckgo_search==8.1.1
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
faiss-cpu==1.13.2
faster-whisper==1.2.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.24.3
Flask==3.1.3
flask-cors==6.0.2
flask-sock==0.7.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kokoro==0.9.4
kokoro-onnx==0.5.0
kubernetes==35.0.0
language-tags==1.2.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
misaki==0.9.4
mmh3==5.2.0
mpmath==1.3.0
mss==10.1.0
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.3
num2words==0.5.14
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.1
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
phonemizer-fork==3.3.2
phue==1.1
pillow==12.1.1
pluggy==1.6.0
posthog==5.4.0
preshed==3.0.12
primp==1.1.2
protobuf==6.33.5
psutil==7.2.2
pvporcupine==4.0.2
PyAudio==0.2.14
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.3.2
pypdf==6.7.5
PyPika==0.51.1
pyproject_hooks==1.2.0
pyserial==3.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
PyYAML==6.0.3
rank-bm25==0.2.2
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segments==2.4.0
sentence-transformers==5.2.3
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tinysegmenter==0.3
tinytuya==1.17.6
tldextract==5.3.1
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wrapt==2.1.2
wsproto==1.3.2
zeroconf==0.148.0
zipp==3.23.0
